requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
beautifulsoup4==4.12.2
Brotli==1.0.9
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==40.0.2
cssselect2==0.7.0
dj-database-url==1.2.0
Django==4.1.4
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-environ==0.9.0
django-heroku==0.3.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-liststyle==0.1
django-paginator==0.2a3
django-paypal==2.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-weasyprint==2.2.0
django-wkhtmltopdf==3.4.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.39.3
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.2
numpy==1.24.3
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.0.1
phonenumbers==8.13.7
phonenumberslite==8.13.7
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pydyf==0.6.0
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pyotp==2.8.0
pypdf==3.8.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
twilio==7.17.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.10
zopfli==0.2.2
